+++ src/media/font/format/dvbf/mod.rs
@@ -1,0 +1,19 @@
+// devela/src/media/font/format/dvbf/mod.rs
+//
+//!
+//
+
+#[cfg(test)]
+mod _test;
+
+mod define; // Dvbf
+mod error; // DvbfError
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            define::*,
+            error::*,
+        };
+    }
+}