+++ src/media/mod.rs
@@ -35,6 +35,7 @@ crate::structural_mods! { // _pub_mods, _crate_internals, _hidden
     _crate_internals {
         pub(crate) use super::_DOC_MEDIA_MODULES;
         pub(crate) use super::{
+            font::_crate_internals::*,
             visual::_crate_internals::*,
         };
     }