+++ docs/changelog.md
@@ -113,7 +113,7 @@
   - impl `ConstInit`.
 
 ### media::font
-- new types: `Fonts`, `FontBitmapView`, `GlyphBitmapView`.
+- new types: `Dvbf`, `Fonts`, `FontBitmapView`, `GlyphBitmapView`.
 - rename `FontBitmap` to `FontBitmapWord`.
 - move standalone font constants to associated `Fonts` constants without `FONT_` prefix.
 - make fonts methods const: `text_<advance|width>`.